refactor: PaymentService 로직을 CommandQuery 분리 패턴으로 리팩토링
